Make time patching by freezegun work with mashumaro
